requirements.txt
pytest
pep8
python-xbrl
xlsxwriter
xlwt
openpyxl
